refactor: moved ShortcutLinksV1 to a dedicated file and refactored dependencies;
